1.0リリース準備：セキュリティ修正・Claude連携の削除・エディタ分割・録画UX改善
## セキュリティ

- ブリッジ(127.0.0.1:14820)のCSRF対策。拡張だけが付けられるX-AuraCap-Bridgeヘッダを必須にし、
  Webページ由来(http/https)のOriginを拒否する。修正前は閲覧中の任意のWebページから画像を注入でき、
  履歴保存・クリップボード上書き・エディタ起動まで通っていた（実機で再現・修正後を検証済み）
- CSPを有効化し、本番用と開発用を分離。カスタムプロトコルはWindows/macOSの両形式を許可
- 履歴系コマンド(edit_history / pin_history / save_recording / discard_recording)が受け取る
  パスを履歴フォルダ配下に限定。任意ファイルの読取・複製・削除を塞ぐ
- ルーペのsizeに上限を設定。無制限だとsize*size*4がu32で溢れ、巨大確保やパニックに至る
- npm audit: postcss(high、パストラバーサル)を解消

## Claude連携の削除

- コード生成をローカルOllama専用にし、画像が外部APIへ出る経路を無くした
- 0.xが平文で保存したAPIキーを、起動時にsettings.jsonから削除する移行処理を追加

## リファクタリング

- Editor.tsx 1515行 → 1008行。editor/{model,raster,Toolbar,ToolIcon,ResultModal} へ分離
- clippy警告5件を解消（clippy / tsc ともにクリーン）

## 録画UXの改善

- 選択確定から録画開始まで枠線が消えていた問題を修正し、待機中も枠を表示する
- 枠の中央に開始までのカウントダウンを表示。数字が録画へ写り込まないよう、非表示から
  画面合成(DWM)に反映されるまでの猶予を待ち時間の内側で消化する
- reccountdown窓をcapabilitiesへ追加。未登録の窓ではAPI呼び出しが無言で拒否される

## 1.0リリース準備

- バージョンを1.0.0に統一（package.json / Cargo.toml / tauri.conf.json / manifest.json）
- LICENSE（プロプライエタリ）と、同梱OSSの表示義務を満たすTHIRD-PARTY-NOTICES.mdを追加
  （生成スクリプト付き。Rust 639クレート＋npm 22パッケージ）
- README更新：macOS対応・録画・OCR・拡張機能・プライバシー方針を追記
- docs/RELEASE.md にリリース手順とコード署名の方針を追加

+++ src-tauri/src/redact.rs
@@ -93,7 +93,7 @@ fn detect(bmp: &[u8]) -> Result<Vec<RedactRegion>, String> {
         .to_rgba8();
 
     let engine = create_engine()?;
-    let max_dim = OcrEngine::MaxImageDimension().map_err(|e| e.to_string())? as u32;
+    let max_dim = OcrEngine::MaxImageDimension().map_err(|e| e.to_string())?;
     if image.width() > max_dim {
         return Err(format!("画像の幅がOCRの上限({max_dim}px)を超えています"));
     }
@@ -179,7 +179,7 @@ fn extract(bmp: &[u8], rect: Option<OcrRect>) -> Result<String, String> {
     }
 
     let engine = create_engine()?;
-    let max_dim = OcrEngine::MaxImageDimension().map_err(|e| e.to_string())? as u32;
+    let max_dim = OcrEngine::MaxImageDimension().map_err(|e| e.to_string())?;
     if image.width() > max_dim {
         return Err(format!("画像の幅がOCRの上限({max_dim}px)を超えています"));
     }
